Demonstrates refactoring into components
